requirements.txt
numpy==1.18.3
opencv-python==4.2.0.34
Pillow==7.1.2
selenium==3.141.0
urllib3==1.25.9
